+++ webknossos/Changelog.md
@@ -26,6 +26,9 @@ For upgrade instructions, please check the respective *Breaking Changes* section
 
 ### Added
 - Added `Dataset.from_images`, which converts images to a Dataset, possibly consisting of multiple layers. [#808](https://github.com/scalableminds/webknossos-libs/pull/808
+- Added `Annotation.open_as_remote_dataset(…)`, which is a shorthand for `Annotation.download(...).get_remote_annotation_dataset()`.
+  [#811](https://github.com/scalableminds/webknossos-libs/pull/811)
+- `skeleton.save()` now also accepts paths with a `.zip` suffix. [#811](https://github.com/scalableminds/webknossos-libs/pull/811)
 
 ### Changed